+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - reworked MIZDB update process
   - now calls a Django management command that checks if an update is available
   - now also downloads a newer version of mizdb.sh if available
+- `mizdb.sh check` now also outputs the results of `docker compose ps`
 
 ## [0.25.1] - 2025-06-27